Split the code up into 3 files
